Start using a map return structure and move username/password validation from model -> handler layer
This is wip, and a checkpoint to document progress. I plan to refactor this later to try a tuple return
structure, and have different validations at both handler and model layers. More on that in a future commit.

Started implementing a 3 key map (:success, :message, :description) to unify return values from validation
functions and db operations. Goal is to simplify passing success/error info back to the handler layer.

Created fn build-result-map to simplify construction of the returned map.
Created fn send-response to send a good/bad response based on the returned map.
Created fn gen-new-user-map to generate a map with all data required to insert a new user into the db.